Refactor RK438Integrator: Merge into RK4Integrator and remove separate file
