+++ tests/integrators/test_rk4.py
@@ -1,4 +1,4 @@
-"""Tests for RK4Integrator."""
+"""Tests for RK4Integrator and RK438Integrator (Kutta's 3/8-rule RK4)."""
 
 import math
 
@@ -6,7 +6,7 @@ import pytest
 import torch
 
 from torchebm.core import GaussianModel, DoubleWellModel
-from torchebm.integrators import RK4Integrator
+from torchebm.integrators import RK4Integrator, RK438Integrator
 from tests.conftest import requires_cuda
 
 
@@ -822,6 +822,834 @@ def test_stiff_ode():
     """Moderately stiff: dx/dt = -50*(x - cos(t))."""
     device = "cuda" if torch.cuda.is_available() else "cpu"
     integrator = RK4Integrator(device=device, dtype=torch.float64)
+
+    x0 = torch.tensor([[1.0]], device=device, dtype=torch.float64)
+    lam = 50.0
+
+    def drift(x_, t_):
+        cos_t = torch.cos(t_).unsqueeze(-1)
+        return -lam * (x_ - cos_t)
+
+    result = integrator.integrate(
+        {"x": x0}, step_size=0.001, n_steps=1000,
+        drift=drift, adaptive=False,
+    )
+
+    expected_approx = math.cos(1.0)
+    assert torch.abs(result["x"].squeeze() - expected_approx).item() < 0.1
+
+
+# ============================================================================
+# RK438Integrator (Kutta's 3/8-rule RK4)
+# ============================================================================
+
+
+################################ Test Fixtures ###########################################
+
+
+@pytest.fixture
+def rk438_integrator():
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    return RK438Integrator(device=device, dtype=torch.float32)
+
+
+@pytest.fixture
+def rk438_integrator_f64():
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    return RK438Integrator(device=device, dtype=torch.float64)
+
+
+################################ Initialization Tests ###########################################
+
+
+def test_rk438_initialization():
+    integrator = RK438Integrator()
+    assert isinstance(integrator, RK438Integrator)
+
+
+def test_rk438_initialization_with_device():
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float32)
+    assert integrator.device == torch.device(device)
+    assert integrator.dtype == torch.float32
+
+
+def test_rk438_no_adaptive_params():
+    """RK4 3/8-rule is fixed-step only — error_weights and order should be None."""
+    integrator = RK438Integrator()
+    assert integrator.error_weights is None
+    assert integrator.order is None
+    assert integrator.fsal is False
+
+
+@requires_cuda
+def test_rk438_cuda():
+    integrator = RK438Integrator(device="cuda", dtype=torch.float32)
+    assert integrator.device == torch.device("cuda")
+
+
+################################ Butcher Tableau Tests ###########################################
+
+
+def test_rk438_tableau_a_values():
+    integrator = RK438Integrator()
+    a = integrator.tableau_a
+
+    assert len(a) == 4
+    assert a[0] == ()
+    assert len(a[1]) == 1
+    assert math.isclose(a[1][0], 1 / 3)
+    assert len(a[2]) == 2
+    assert math.isclose(a[2][0], -1 / 3)
+    assert math.isclose(a[2][1], 1.0)
+    assert len(a[3]) == 3
+    assert math.isclose(a[3][0], 1.0)
+    assert math.isclose(a[3][1], -1.0)
+    assert math.isclose(a[3][2], 1.0)
+
+
+def test_rk438_tableau_b_values():
+    integrator = RK438Integrator()
+    b = integrator.tableau_b
+
+    assert len(b) == 4
+    assert math.isclose(b[0], 1 / 8)
+    assert math.isclose(b[1], 3 / 8)
+    assert math.isclose(b[2], 3 / 8)
+    assert math.isclose(b[3], 1 / 8)
+
+
+def test_rk438_tableau_c_values():
+    integrator = RK438Integrator()
+    c = integrator.tableau_c
+
+    assert len(c) == 4
+    assert c[0] == 0.0
+    assert math.isclose(c[1], 1 / 3)
+    assert math.isclose(c[2], 2 / 3)
+    assert math.isclose(c[3], 1.0)
+
+
+def test_rk438_n_stages():
+    integrator = RK438Integrator()
+    assert integrator.n_stages == 4
+
+
+def test_rk438_tableau_consistency_b_sum():
+    integrator = RK438Integrator()
+    b = integrator.tableau_b
+    assert math.isclose(sum(b), 1.0, abs_tol=1e-14)
+
+
+def test_rk438_tableau_consistency_row_sums():
+    integrator = RK438Integrator()
+    a = integrator.tableau_a
+    c = integrator.tableau_c
+
+    for i in range(len(a)):
+        row_sum = sum(a[i]) if a[i] else 0.0
+        assert math.isclose(row_sum, c[i], abs_tol=1e-14), (
+            f"Row {i}: sum(a[{i}]) = {row_sum} != c[{i}] = {c[i]}"
+        )
+
+
+################################ Step Method Tests ###########################################
+
+
+def test_rk438_step_ode(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    state = {"x": x}
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step(state, step_size=0.01, drift=drift)
+
+    assert "x" in result
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_step_with_explicit_time(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    t = torch.ones(10, device=device) * 0.5
+    drift = lambda x_, t_: -x_ * (1 + t_[:, None])
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift, t=t)
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_step_with_model_gradient(rk438_integrator, gaussian_model):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    drift = lambda x_, t_: -gaussian_model.gradient(x_)
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift)
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_step_requires_drift(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+
+    with pytest.raises(ValueError, match="drift must be provided explicitly"):
+        rk438_integrator.step({"x": x}, step_size=0.01)
+
+
+def test_rk438_step_default_time_is_zero(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.ones(10, 2, device=device)
+    drift_calls = []
+
+    def recording_drift(x_, t_):
+        drift_calls.append(t_.clone())
+        return -x_
+
+    rk438_integrator.step({"x": x}, step_size=0.01, drift=recording_drift)
+
+    assert len(drift_calls) > 0
+    assert torch.allclose(drift_calls[0], torch.zeros(10, device=device))
+
+
+def test_rk438_step_scalar_step_size(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=lambda x_, t_: -x_)
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_step_tensor_step_size(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    step_size = torch.tensor(0.01, device=device)
+
+    result = rk438_integrator.step({"x": x}, step_size=step_size, drift=lambda x_, t_: -x_)
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+################################ Integrate Method Tests ###########################################
+
+
+def test_rk438_integrate_ode(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.ones(10, 2, device=device)
+    drift = lambda x_, t_: -x_
+    n_steps = 50
+    step_size = 0.02
+
+    result = rk438_integrator.integrate(
+        {"x": x}, step_size=step_size, n_steps=n_steps,
+        drift=drift, adaptive=False,
+    )
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+    expected = x * math.exp(-1.0)
+    assert torch.allclose(result["x"], expected, rtol=0.001)
+
+
+def test_rk438_integrate_with_time_grid(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.ones(10, 2, device=device)
+    drift = lambda x_, t_: -x_
+    t = torch.linspace(0, 1, 51, device=device)
+
+    result = rk438_integrator.integrate(
+        {"x": x}, step_size=0.02, n_steps=50,
+        drift=drift, t=t, adaptive=False,
+    )
+
+    expected = x * math.exp(-1.0)
+    assert torch.allclose(result["x"], expected, rtol=0.001)
+
+
+def test_rk438_integrate_with_model(rk438_integrator, gaussian_model):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device) * 3
+    drift = lambda x_, t_: -gaussian_model.gradient(x_)
+
+    result = rk438_integrator.integrate(
+        {"x": x}, step_size=0.01, n_steps=100,
+        drift=drift, adaptive=False,
+    )
+
+    initial_dist = torch.norm(x, dim=-1).mean()
+    final_dist = torch.norm(result["x"], dim=-1).mean()
+    assert final_dist < initial_dist
+
+
+def test_rk438_integrate_invalid_n_steps(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    drift = lambda x_, t_: -x_
+
+    with pytest.raises(ValueError, match="n_steps must be positive"):
+        rk438_integrator.integrate(
+            {"x": x}, step_size=0.01, n_steps=0,
+            drift=drift, adaptive=False,
+        )
+
+    with pytest.raises(ValueError, match="n_steps must be positive"):
+        rk438_integrator.integrate(
+            {"x": x}, step_size=0.01, n_steps=-5,
+            drift=drift, adaptive=False,
+        )
+
+
+def test_rk438_integrate_time_grid_validation(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    drift = lambda x_, t_: -x_
+
+    t_wrong_shape = torch.linspace(0, 1, 50, device=device).reshape(5, 10)
+    with pytest.raises(ValueError, match="t must be a 1D tensor"):
+        rk438_integrator.integrate(
+            {"x": x}, step_size=0.02, n_steps=50,
+            drift=drift, t=t_wrong_shape, adaptive=False,
+        )
+
+    t_too_short = torch.tensor([0.0], device=device)
+    with pytest.raises(ValueError, match="t must be a 1D tensor with length >= 2"):
+        rk438_integrator.integrate(
+            {"x": x}, step_size=0.02, n_steps=1,
+            drift=drift, t=t_too_short, adaptive=False,
+        )
+
+
+def test_rk438_single_step_integration_matches_step(rk438_integrator_f64):
+    """1-step integrate should match a single step() call."""
+    device = rk438_integrator_f64.device
+    x = torch.randn(10, 2, device=device, dtype=torch.float64)
+    drift = lambda x_, t_: -x_
+
+    step_result = rk438_integrator_f64.step(
+        {"x": x.clone()}, step_size=0.1, drift=drift,
+    )
+    integrate_result = rk438_integrator_f64.integrate(
+        {"x": x.clone()}, step_size=0.1, n_steps=1,
+        drift=drift, adaptive=False,
+    )
+
+    assert torch.allclose(step_result["x"], integrate_result["x"], atol=1e-12)
+
+
+def test_rk438_no_adaptive_mode():
+    """RK4 3/8-rule has no error_weights, so fixed-step is the reliable path."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+    x = torch.ones(5, 2, device=device, dtype=torch.float64)
+    drift = lambda x_, t_: -x_
+
+    result = integrator.integrate(
+        {"x": x}, step_size=0.02, n_steps=50,
+        drift=drift, adaptive=False,
+    )
+    expected = x * math.exp(-1.0)
+    assert torch.allclose(result["x"], expected, rtol=0.001)
+
+
+################################ Manual Verification Tests ###########################################
+
+
+def test_manual_rk438_step_linear_decay():
+    r"""Manual verification of a single 3/8-rule RK4 step for f(x,t)=-x.
+
+    Uses the tableau a=((), (1/3,), (-1/3, 1), (1, -1, 1)), b=(1/8, 3/8, 3/8, 1/8).
+    """
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x = torch.tensor([[1.0, 2.0]], device=device, dtype=torch.float64)
+    drift = lambda x_, t_: -x_
+    h = 0.1
+
+    a = integrator.tableau_a
+    b = integrator.tableau_b
+
+    k = []
+    for i in range(4):
+        if i == 0:
+            x_stage = x
+        else:
+            dx = torch.zeros_like(x)
+            for j in range(i):
+                if a[i][j] != 0:
+                    dx = dx + a[i][j] * k[j]
+            x_stage = x + h * dx
+        k.append(-x_stage)
+
+    x_expected = x.clone()
+    for i in range(4):
+        x_expected = x_expected + h * b[i] * k[i]
+
+    result = integrator.step({"x": x}, step_size=h, drift=drift)
+
+    assert torch.allclose(result["x"], x_expected, atol=1e-14)
+
+
+def test_manual_rk438_step_constant_drift():
+    """For constant drift f(x,t)=c, all stages equal c, x_new = x + h*c."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x = torch.tensor([[1.0, 2.0]], device=device, dtype=torch.float64)
+    c = torch.tensor([[3.0, -1.0]], device=device, dtype=torch.float64)
+    drift = lambda x_, t_: c
+    h = 0.5
+
+    result = integrator.step({"x": x}, step_size=h, drift=drift)
+
+    expected = x + h * c
+    assert torch.allclose(result["x"], expected, atol=1e-14)
+
+
+def test_manual_rk438_step_linear_time_drift():
+    r"""For f(x,t)=t, x(0)=0, RK4 should integrate x(h)=h²/2 exactly.
+
+    4th-order methods integrate polynomials up to degree 3 exactly.
+    f(x,t)=t ⇒ x(h) = h²/2.
+    """
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x = torch.tensor([[0.0, 0.0]], device=device, dtype=torch.float64)
+    drift = lambda x_, t_: t_.unsqueeze(-1).expand_as(x_)
+    h = 1.0
+
+    result = integrator.step(
+        {"x": x}, step_size=h, drift=drift,
+        t=torch.zeros(1, device=device, dtype=torch.float64),
+    )
+
+    expected = x + h**2 / 2
+    assert torch.allclose(result["x"], expected, atol=1e-12)
+
+
+def test_manual_rk438_cubic_time_drift():
+    r"""For f(x,t)=t³, x(0)=0, solution is x(h)=h⁴/4.
+
+    A 4th-order RK method integrates a degree-3 polynomial ODE exactly:
+    dx/dt = t³ ⇒ x(1) = 1/4.
+    """
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x0 = torch.zeros(1, 1, device=device, dtype=torch.float64)
+    drift = lambda x_, t_: (t_**3).unsqueeze(-1).expand_as(x_)
+
+    result = integrator.step(
+        {"x": x0}, step_size=1.0, drift=drift,
+        t=torch.zeros(1, device=device, dtype=torch.float64),
+    )
+
+    expected = torch.tensor([[0.25]], device=device, dtype=torch.float64)
+    assert torch.allclose(result["x"], expected, atol=1e-12)
+
+
+def test_manual_rk438_quadratic_drift():
+    """For f(x,t)=t², x(0)=0, solution x(1)=1/3. 4th-order RK integrates exactly."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x0 = torch.zeros(1, 1, device=device, dtype=torch.float64)
+    drift = lambda x_, t_: (t_**2).unsqueeze(-1).expand_as(x_)
+
+    result = integrator.step(
+        {"x": x0}, step_size=1.0, drift=drift,
+        t=torch.zeros(1, device=device, dtype=torch.float64),
+    )
+
+    expected = torch.tensor([[1.0 / 3.0]], device=device, dtype=torch.float64)
+    assert torch.allclose(result["x"], expected, atol=1e-14)
+
+
+def test_rk438_exponential_decay_numerical():
+    r"""Numerical check: 3/8-rule RK4 approximation to e^{-h}.
+
+    For f(x,t)=-x, h=0.1, x(0)=1, the 3/8-rule gives a 4th-order accurate
+    approximation to exp(-0.1) = 0.90483741803...
+    """
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x = torch.tensor([[1.0]], device=device, dtype=torch.float64)
+    drift = lambda x_, t_: -x_
+    h = 0.1
+
+    result = integrator.step({"x": x}, step_size=h, drift=drift)
+
+    assert math.isclose(result["x"].item(), math.exp(-h), rel_tol=1e-5)
+
+
+################################ Convergence and Accuracy Tests ###########################################
+
+
+def test_rk438_fourth_order_convergence():
+    """Halving step size should reduce error by ~2^4 = 16x."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x0 = torch.tensor([[2.0, 3.0]], device=device, dtype=torch.float64)
+    drift = lambda x_, t_: -x_
+
+    step_sizes = [0.1, 0.05, 0.025]
+    errors = []
+
+    for step_size in step_sizes:
+        n_steps = int(1.0 / step_size)
+        result = integrator.integrate(
+            {"x": x0.clone()}, step_size=step_size, n_steps=n_steps,
+            drift=drift, adaptive=False,
+        )
+        expected = x0 * math.exp(-1.0)
+        error = torch.abs(result["x"] - expected).max().item()
+        errors.append(error)
+
+    for i in range(len(errors) - 1):
+        ratio = errors[i] / errors[i + 1]
+        assert ratio > 12, f"Expected ~16x error reduction for 4th order, got {ratio:.1f}"
+
+
+def test_rk438_vs_heun_accuracy():
+    """3/8-rule RK4 (order 4) should be more accurate than Heun (order 2)."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    dtype = torch.float64
+    from torchebm.integrators import HeunIntegrator
+
+    rk438 = RK438Integrator(device=device, dtype=dtype)
+    heun = HeunIntegrator(device=device, dtype=dtype)
+
+    x0 = torch.tensor([[1.0, 1.0]], device=device, dtype=dtype)
+    drift = lambda x_, t_: -x_
+    step_size = 0.1
+    n_steps = 10
+
+    rk438_result = rk438.integrate(
+        {"x": x0.clone()}, step_size=step_size, n_steps=n_steps,
+        drift=drift, adaptive=False,
+    )
+    heun_result = heun.integrate(
+        {"x": x0.clone()}, step_size=step_size, n_steps=n_steps,
+        drift=drift, adaptive=False,
+    )
+
+    expected = x0 * math.exp(-1.0)
+    rk438_error = torch.abs(rk438_result["x"] - expected).max().item()
+    heun_error = torch.abs(heun_result["x"] - expected).max().item()
+
+    assert rk438_error < heun_error, (
+        f"3/8-rule RK4 error {rk438_error} should be less than Heun error {heun_error}"
+    )
+
+
+def test_rk438_vs_rk4_same_order():
+    """3/8-rule and classical RK4 are both 4th order — their solutions should
+    agree to roughly O(h^5) on smooth problems."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    dtype = torch.float64
+
+    rk438 = RK438Integrator(device=device, dtype=dtype)
+    rk4 = RK4Integrator(device=device, dtype=dtype)
+
+    x0 = torch.tensor([[1.0, 1.0]], device=device, dtype=dtype)
+    drift = lambda x_, t_: -x_
+    step_size = 0.01
+    n_steps = 100
+
+    rk438_result = rk438.integrate(
+        {"x": x0.clone()}, step_size=step_size, n_steps=n_steps,
+        drift=drift, adaptive=False,
+    )
+    rk4_result = rk4.integrate(
+        {"x": x0.clone()}, step_size=step_size, n_steps=n_steps,
+        drift=drift, adaptive=False,
+    )
+
+    assert torch.allclose(rk438_result["x"], rk4_result["x"], atol=1e-8)
+
+
+def test_rk438_exponential_decay_convergence():
+    """Errors should decrease monotonically as step size shrinks."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x0 = torch.tensor([[2.0, 3.0]], device=device, dtype=torch.float64)
+    drift = lambda x_, t_: -x_
+
+    step_sizes = [0.1, 0.01, 0.001]
+    errors = []
+
+    for step_size in step_sizes:
+        n_steps = int(1.0 / step_size)
+        result = integrator.integrate(
+            {"x": x0.clone()}, step_size=step_size, n_steps=n_steps,
+            drift=drift, adaptive=False,
+        )
+        expected = x0 * math.exp(-1.0)
+        error = torch.abs(result["x"] - expected).max().item()
+        errors.append(error)
+
+    for i in range(len(errors) - 1):
+        assert errors[i + 1] < errors[i]
+
+
+def test_rk438_nonlinear_drift_cubic():
+    r"""dx/dt = -x³, x(0)=0.5, solution x(t)=1/√(2t+1/x₀²). At t=1: x=1/√6."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x = torch.ones(10, 2, device=device, dtype=torch.float64) * 0.5
+    drift = lambda x_, t_: -(x_**3)
+
+    result = integrator.integrate(
+        {"x": x}, step_size=0.01, n_steps=100,
+        drift=drift, adaptive=False,
+    )
+
+    x0_val = 0.5
+    expected_val = 1.0 / math.sqrt(2 * 1.0 + 1.0 / x0_val**2)
+    expected = torch.full_like(x, expected_val)
+
+    assert torch.allclose(result["x"], expected, atol=1e-8)
+
+
+def test_rk438_harmonic_oscillator():
+    """dx/dt=v, dv/dt=-x. After 2π the state returns to start."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    state_vec = torch.tensor([[1.0, 0.0]], device=device, dtype=torch.float64)
+
+    def drift(s, t_):
+        return torch.cat([s[:, 1:2], -s[:, 0:1]], dim=-1)
+
+    t_final = 2 * math.pi
+    n_steps = 1000
+    step_size = t_final / n_steps
+    result = integrator.integrate(
+        {"x": state_vec}, step_size=step_size, n_steps=n_steps,
+        drift=drift, adaptive=False,
+    )
+
+    assert torch.allclose(result["x"], state_vec, atol=1e-5)
+
+
+def test_rk438_polynomial_degree4_not_exact_single_step():
+    r"""dx/dt = t⁴, x(0)=0 ⇒ x(1)=1/5. A 4th-order method should NOT
+    integrate t⁴ exactly in a single step (only exact up to degree 3)."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+
+    x0 = torch.zeros(1, 1, device=device, dtype=torch.float64)
+    drift = lambda x_, t_: (t_**4).unsqueeze(-1).expand_as(x_)
+
+    result = integrator.step(
+        {"x": x0}, step_size=1.0, drift=drift,
+        t=torch.zeros(1, device=device, dtype=torch.float64),
+    )
+
+    exact = torch.tensor([[0.2]], device=device, dtype=torch.float64)
+    assert not torch.allclose(result["x"], exact, atol=1e-10)
+    assert torch.allclose(result["x"], exact, atol=0.02)
+
+
+################################ Reproducibility Tests ###########################################
+
+
+def test_rk438_reproducibility_fixed(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    drift = lambda x_, t_: -x_
+
+    result1 = rk438_integrator.integrate(
+        {"x": x.clone()}, step_size=0.01, n_steps=100,
+        drift=drift, adaptive=False,
+    )
+    result2 = rk438_integrator.integrate(
+        {"x": x.clone()}, step_size=0.01, n_steps=100,
+        drift=drift, adaptive=False,
+    )
+
+    assert torch.allclose(result1["x"], result2["x"])
+
+
+################################ Edge Cases ###########################################
+
+
+def test_rk438_large_batch_size(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(1000, 10, device=device)
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift)
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_high_dimension(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 100, device=device)
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift)
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_small_step_size(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step({"x": x}, step_size=1e-8, drift=drift)
+
+    assert torch.all(torch.isfinite(result["x"]))
+    assert torch.allclose(result["x"], x, atol=1e-5)
+
+
+def test_rk438_large_step_size(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.ones(10, 2, device=device) * 0.1
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step({"x": x}, step_size=1.0, drift=drift)
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_large_values(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device) * 1000
+    drift = lambda x_, t_: -x_ * 0.001
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift)
+
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_zero_drift(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    drift = lambda x_, t_: torch.zeros_like(x_)
+
+    result = rk438_integrator.step({"x": x}, step_size=0.1, drift=drift)
+
+    assert torch.allclose(result["x"], x, atol=1e-7)
+
+
+def test_rk438_zero_step_size(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(10, 2, device=device)
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step({"x": x}, step_size=0.0, drift=drift)
+
+    assert torch.allclose(result["x"], x, atol=1e-7)
+
+
+def test_rk438_time_dependent_drift(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.ones(10, 2, device=device)
+    t = torch.ones(10, device=device) * 0.5
+    drift = lambda x_, t_: -x_ * (1 + t_[:, None])
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift, t=t)
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_multidimensional_batch(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(8, 4, 3, device=device)
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift)
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_single_sample(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(1, 2, device=device)
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift)
+
+    assert result["x"].shape == x.shape
+    assert torch.all(torch.isfinite(result["x"]))
+
+
+def test_rk438_dtype_float32(rk438_integrator):
+    device = rk438_integrator.device
+    x = torch.randn(5, 2, device=device, dtype=torch.float32)
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator.step({"x": x}, step_size=0.01, drift=drift)
+    assert result["x"].dtype == torch.float32
+
+
+def test_rk438_dtype_float64(rk438_integrator_f64):
+    device = rk438_integrator_f64.device
+    x = torch.randn(5, 2, device=device, dtype=torch.float64)
+    drift = lambda x_, t_: -x_
+
+    result = rk438_integrator_f64.step({"x": x}, step_size=0.01, drift=drift)
+    assert result["x"].dtype == torch.float64
+
+
+def test_rk438_double_well_convergence():
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
+    model = DoubleWellModel(barrier_height=2.0).to(device)
+
+    x = torch.ones(10, 1, device=device, dtype=torch.float64) * 1.5
+    drift = lambda x_, t_: -model.gradient(x_.float()).double()
+
+    result = integrator.integrate(
+        {"x": x}, step_size=0.01, n_steps=200,
+        drift=drift, adaptive=False,
+    )
+
+    assert torch.all(torch.isfinite(result["x"]))
+    assert torch.all(torch.abs(result["x"]) > 0.5)
+
+
+def test_rk438_quadratic_drift():
+    """f(x)=-x², starting from 1 all values should decrease."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float32)
+
+    x = torch.ones(10, 2, device=device)
+    drift = lambda x_, t_: -(x_**2)
+    n_steps = 50
+    t = torch.linspace(0, 0.5, n_steps, device=device)
+
+    result = integrator.integrate(
+        {"x": x}, step_size=0.01, n_steps=n_steps,
+        drift=drift, t=t, adaptive=False,
+    )
+
+    assert torch.all(torch.isfinite(result["x"]))
+    assert torch.all(result["x"] < x)
+
+
+def test_rk438_stiff_ode():
+    """Moderately stiff: dx/dt = -50*(x - cos(t))."""
+    device = "cuda" if torch.cuda.is_available() else "cpu"
+    integrator = RK438Integrator(device=device, dtype=torch.float64)
 
     x0 = torch.tensor([[1.0]], device=device, dtype=torch.float64)
     lam = 50.0
